requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.2.7
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
cloudinary==1.36.0
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.14.6
datedelta==1.4
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
docopt==0.6.2
emoji==1.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.103.2
filelock==3.13.1
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.26
flex==6.14.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
fuzzywuzzy==0.18.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
grapheme==0.6.0
graphlib-backport==1.0.3
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
identify==2.5.31
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.13.0
InstructorEmbedding==1.0.1
ipython==8.12.0
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-spec==0.1.6
kaggle==1.5.16
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.333
langchain-experimental==0.0.39
langsmith==0.0.63
lazy-object-proxy==1.9.0
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mo-dots==9.401.23144
mo-future==7.401.23144
mo-imports==7.401.23144
mo-parsing==8.401.23144
mo-sql-parsing==9.402.23144
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.16.1
openai==1.1.1
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.6
openpyxl==2.4.11
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pathable==0.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
posthog==3.0.2
prance==0.22.11.4.0
pre-commit==3.5.0
prompt-toolkit==3.0.39
protobuf==4.25.0
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycharts==0.1.5
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pymongo==4.3.3
pyparsing==3.1.1
PyPika==0.48.9
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
recognizers-text-sequence==1.0.2a2
recognizers-text-suite==1.0.2a2
records==0.5.3
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlparse==0.4.2
stack-data==0.6.3
starlette==0.27.0
strict-rfc3339==0.7
sympy==1.12
tablib==3.5.0
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.2
traitlets==5.13.0
transformers==4.35.0
typer==0.9.0
types-requests==0.1.13
typing-inspect==0.8.0
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
validate-email==1.3
virtualenv==20.24.6
watchfiles==0.21.0
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
